# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "description": "This is an example API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/subOption": {
            "get": {
                "tags": [
                    "SubOption"
                ],
                "summary": "Get all SubOption",
                "operationId": "b5ab787afb14581db816b18350dae795",
                "responses": {
                    "200": {
                        "description": "Everything OK"
                    },
                    "403": {
                        "description": "Access Denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/subOption/store": {
            "post": {
                "tags": [
                    "SubOption"
                ],
                "description": "",
                "operationId": "d40100de79848801086ab82fa6b695a8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "option_id": {
                                                "type": "integer"
                                            },
                                            "name_ar": {
                                                "type": "string"
                                            },
                                            "name_en": {
                                                "type": "string"
                                            },
                                            "value": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "double"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "option_id": "option_id",
                                    "name_ar": "name_ar",
                                    "name_en": "name_en",
                                    "value": "value",
                                    "price": "price"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "option_id": {
                                            "type": "integr",
                                            "example": "option_id"
                                        },
                                        "name_ar": {
                                            "type": "string",
                                            "example": "name_ar"
                                        },
                                        "name_en": {
                                            "type": "string",
                                            "example": "name_en"
                                        },
                                        "value": {
                                            "type": "string",
                                            "example": "value"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "double",
                                            "example": "price"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "CredentialsError = 1;Required = 100;MaxLength255 = 101;Unique = 103;"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/subOption/show/{id}": {
            "get": {
                "tags": [
                    "SubOption"
                ],
                "summary": "Get  SubOption",
                "operationId": "cff8d067aba12e423c438470ba0ff3d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything OK"
                    },
                    "403": {
                        "description": "Access Denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/subOption/update/{id}": {
            "post": {
                "tags": [
                    "SubOption"
                ],
                "description": "",
                "operationId": "f9575caca3164a78b7d505a1e46280f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "option_id": {
                                                "type": "integer"
                                            },
                                            "name_ar": {
                                                "type": "string"
                                            },
                                            "name_en": {
                                                "type": "string"
                                            },
                                            "value": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "double"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "option_id": "option_id",
                                    "name_ar": "name_ar",
                                    "name_en": "name_en",
                                    "value": "value",
                                    "price": "price"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "option_id": {
                                            "type": "integr",
                                            "example": "option_id"
                                        },
                                        "name_ar": {
                                            "type": "string",
                                            "example": "name_ar"
                                        },
                                        "name_en": {
                                            "type": "string",
                                            "example": "name_en"
                                        },
                                        "value": {
                                            "type": "string",
                                            "example": "value"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "double",
                                            "example": "price"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "CredentialsError = 1;Required = 100;MaxLength255 = 101;Unique = 103;"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/subOption/delete/{id}": {
            "delete": {
                "tags": [
                    "SubOption"
                ],
                "summary": "delete  SubOption",
                "operationId": "17aa3340f9eaa42191243385e0764996",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything OK"
                    },
                    "403": {
                        "description": "Access Denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}